refactor: move user use cases to its own folder
